WIP: decouple PersonalDetailsReviewPresenter
