+++ app/components/personal_details_review_component.rb
@@ -3,15 +3,10 @@ class PersonalDetailsReviewComponent < ActionView::Component::Base
 
   def initialize(application_form:)
     @application_form = application_form
-    @personal_details_form = CandidateInterface::PersonalDetailsForm.build_from_application(
-      application_form,
-    )
   end
 
   def rows
-    CandidateInterface::PersonalDetailsReviewPresenter
-      .new(@personal_details_form)
-      .rows
+    PersonalDetailsReviewPresenter.new(application_form).rows
   end
 
 private
